Move shared logic from HomePage implementations to separate file
