+++ speedy-recovery-client/src/service/DateUtils.js
@@ -1,0 +1,6 @@
+export function formatDate(date) {
+  if (date) {
+    return new Date(date).toLocaleString("en-uk");
+  }
+  return "Invalid date";
+}
